Seismic design, structures, and income inequality in Mexico City
The Python code is for plotting the seismic response spectrum based on the 2015 Handbook of Civil Works of Federal Commission of Electricity Seismic Design Chapter for Mexico. The R code is for relating 2018 trimestral income brackets in CDMX to terrain zones (bland, mixed, firm). This excel file was made through an ArcGIS analysis.